[面試經驗] 拆分 Checkbox private (#1200)
title:	[面試經驗] 拆分 Checkbox private
state:	OPEN
author:	peteranny
labels:	
assignees:	
reviewers:	barry800414 (Approved)
projects:	
milestone:	
number:	1200
url:	https://github.com/goodjoblife/GoodJobShare/pull/1200
additions:	219
deletions:	196
auto-merge:	disabled
--
#886 

## 這個 PR 是？ <!-- 必填 -->

Checkbox/private 裡包含許多元件，這個PR把他們分開來，以利之後的擴充

## 我應該如何手動測試？ <!-- 必填 -->

- [ ] http://localhost:3000/share 面試表單應不受影響
