+++ src/components/common/FormBuilder/QuestionBuilder/Checkbox/private/utils.js
@@ -1,0 +1,9 @@
+import R from 'ramda';
+
+export const toggle = (value, values) => {
+  if (R.contains(value, values)) {
+    return R.without(value, values);
+  } else {
+    return R.append(value, values);
+  }
+};
